ok, in Darwin tv_usec is 'int' and in Linux it can be 'long' - here is the compromise

+++ source/stream.cpp
@@ -346,6 +346,6 @@ CStream& CStream::Element(const std::string& crsName, double data)
 CStream& CStream::Element(const std::string& crsName, timeval data)
   {
   char ac[64];
-  snprintf (ac, sizeof(ac), "%ld:%ld", data.tv_sec, data.tv_usec);
+  snprintf (ac, sizeof(ac), "%ld:%ld", data.tv_sec, (long)data.tv_usec);
   return Element(crsName, std::string(ac));
   } //
